fix: hog file with correct file
